+++ src/lib/output/themes/default/layouts/default.tsx
@@ -28,10 +28,10 @@ export const defaultLayout = (
             {context.options.getValue("customCss") && (
                 <link rel="stylesheet" href={context.relativeURL("assets/custom.css", true)} />
             )}
+            <script defer src={context.relativeURL("assets/main.js", true)}></script>
             {context.options.getValue("customJs") && (
                 <script defer src={context.relativeURL("assets/custom.js", true)}></script>
             )}
-            <script defer src={context.relativeURL("assets/main.js", true)}></script>
             <script async src={context.relativeURL("assets/icons.js", true)} id="tsd-icons-script"></script>
             <script async src={context.relativeURL("assets/search.js", true)} id="tsd-search-script"></script>
             <script async src={context.relativeURL("assets/navigation.js", true)} id="tsd-nav-script"></script>